Resolve "Provide more server components on web"

Closes #39

See merge request andriy.berzhavych/sunnydaysapp!40

+++ server/src/modules/cities/cities.service.ts
@@ -1,6 +1,15 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DeleteResult, Repository } from 'typeorm';
+
+import {
+  NO_MATCHING_LOCATION_FOUND_ERROR_CODE,
+  WeatherManagementService,
+} from '@modules/weather-management';
 
 import { City } from './entities';
 
@@ -8,17 +17,49 @@ import { City } from './entities';
 export class CitiesService {
   constructor(
     @InjectRepository(City) private readonly citiesRepository: Repository<City>,
+    private readonly weatherManagementService: WeatherManagementService,
   ) {}
 
-  async createCity(cityName: string): Promise<City> {
-    let city = await this.findByName(cityName);
+  async createCity(
+    cityName: string,
+    forecastDaysAmount: number,
+  ): Promise<City> {
+    try {
+      const response = await this.weatherManagementService.getCityWeather(
+        cityName,
+        forecastDaysAmount,
+      );
 
-    if (city) {
-      return city;
+      const forecast =
+        this.weatherManagementService.mapResponseToWeatherForecast(
+          response,
+          cityName,
+        );
+
+      await this.weatherManagementService.cacheForecast(forecast);
+
+      let city = await this.findByName(cityName);
+
+      if (city) {
+        return city;
+      }
+
+      city = this.citiesRepository.create({ name: cityName.toLowerCase() });
+      return this.citiesRepository.save(city);
+    } catch (error) {
+      if (
+        error.response &&
+        error.response.data.error &&
+        error.response.data.error.code === NO_MATCHING_LOCATION_FOUND_ERROR_CODE
+      ) {
+        throw new BadRequestException(
+          error.response.data.error.message,
+          error.response.status,
+        );
+      } else {
+        throw error;
+      }
     }
-
-    city = this.citiesRepository.create({ name: cityName.toLowerCase() });
-    return this.citiesRepository.save(city);
   }
 
   async findByName(name: string): Promise<City> {
@@ -33,6 +74,11 @@ export class CitiesService {
 
   async deleteCity(name: string): Promise<DeleteResult> {
     const city = await this.findByName(name);
+
+    if (!city) {
+      throw new NotFoundException(`City '${name}' not found!`);
+    }
+
     return this.citiesRepository.delete(city.id);
   }
 }
